Improve ergonomics of client write error handling.
This is a farily large changeset, and is also a breaking change.
This introduces an `AppError` type (which actually existed in older
versions of this crate). This allows for the Raft system to
differentiate between fatal storage errors which should cause the system
to shutdown vs errors which should be propagated back to the client for
application specific error handling.

Overall, this constitutes a nice improvement to the RaftStorage
interface, as a flat result type may be used for propagating errors. No
result nesting required.

Added tests to cover cases where AppError is returned from the
`RaftStorage::apply_entry_to_state_machine` handler.